#13 Refactor JWTAuthenticationFilter
