+++ src/main/kotlin/io/github/raeperd/realworldspringbootkotlin/web/jwt/JWTPayloadArgumentResolver.kt
@@ -1,7 +1,7 @@
-package io.github.raeperd.realworldspringbootkotlin.web
+package io.github.raeperd.realworldspringbootkotlin.web.jwt
 
 import io.github.raeperd.realworldspringbootkotlin.domain.JWTPayload
-import io.github.raeperd.realworldspringbootkotlin.web.JWTAuthenticationFilter.Companion.JWT_PAYLOAD_ATTRIBUTE_NAME
+import io.github.raeperd.realworldspringbootkotlin.web.jwt.JWTAuthenticationFilter.Companion.JWT_PAYLOAD_ATTRIBUTE_NAME
 import org.springframework.core.MethodParameter
 import org.springframework.web.bind.support.WebDataBinderFactory
 import org.springframework.web.context.request.NativeWebRequest
